+++ src/main/kotlin/com/pestphp/pest/indexers/PestTestIndex.kt
@@ -6,29 +6,33 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.util.indexing.DataIndexer
 import com.intellij.util.indexing.DefaultFileTypeSpecificInputFilter
 import com.intellij.util.indexing.FileBasedIndex
+import com.intellij.util.indexing.FileBasedIndexExtension
 import com.intellij.util.indexing.FileContent
 import com.intellij.util.indexing.ID
-import com.intellij.util.indexing.ScalarIndexExtension
+import com.intellij.util.io.DataExternalizer
 import com.intellij.util.io.EnumeratorStringDescriptor
 import com.intellij.util.io.KeyDescriptor
 import com.jetbrains.php.lang.PhpFileType
+import com.jetbrains.php.lang.psi.stubs.indexes.StringSetDataExternalizer
+import com.pestphp.pest.getPestTestName
+import com.pestphp.pest.getPestTests
 import com.pestphp.pest.isPestTestFile
 import gnu.trove.THashMap
 
-class PestTestIndex : ScalarIndexExtension<String>() {
-    override fun getName(): ID<String, Void> {
+class PestTestIndex : FileBasedIndexExtension<String, Set<String>>() {
+    override fun getName(): ID<String, Set<String>> {
         return key
     }
 
     override fun getVersion(): Int {
-        return 0
+        return 2
     }
 
     override fun dependsOnFileContent(): Boolean {
         return true
     }
 
-    override fun getIndexer(): DataIndexer<String, Void, FileContent> {
+    override fun getIndexer(): DataIndexer<String, Set<String>, FileContent> {
         return DataIndexer { inputData ->
             val file = inputData.psiFile
 
@@ -36,8 +40,10 @@ class PestTestIndex : ScalarIndexExtension<String>() {
                 return@DataIndexer mapOf()
             }
 
-            val map = THashMap<String, Void>()
-            map[file.name] = null
+            val map = THashMap<String, Set<String>>()
+            map[file.name] = file.getPestTests()
+                .mapNotNull { it.getPestTestName() }
+                .toSet()
             return@DataIndexer map
         }
     }
@@ -60,7 +66,11 @@ class PestTestIndex : ScalarIndexExtension<String>() {
         return EnumeratorStringDescriptor.INSTANCE
     }
 
+    override fun getValueExternalizer(): DataExternalizer<Set<String>> {
+        return StringSetDataExternalizer.INSTANCE
+    }
+
     companion object {
-        val key = ID.create<String, Void>("php.pest")
+        val key = ID.create<String, Set<String>>("php.pest")
     }
 }